Uppdaterat Formatter
Lagt in Formatter i Utilities
Lagt till employee.add() som saknades
